[MOVE] checkPermissoin into loadNodeWithPermission
