+++ server/api/middleware/loadNodeWithPermssion.js
@@ -10,8 +10,6 @@ const ObjectId = mongoose.Types.ObjectId;
 const Node = require("../models/nodeModel");
 const User = require("../models/userModel");
 
-const checkPermission = require("./checkPermission")
-
 const errorNotFound = { message: "No valid node found for provided ID", code: 404 }
 const errorObjectId = { message: "Provided ObjectId is not valid.", code : 400 };
 const errorInvalidType = { message: "Provided action type is not valid", code: 500 };
@@ -77,6 +75,46 @@ function load(req) {
         }
     });
 }
+
+function checkPermission (req, user) {
+    let permissionList = new Set();
+
+    if (!user) {
+        // If guest, add guest filter
+        permissionList.add('defaultGuest')
+    } else {
+        // Signuped user who has special permissions
+        if (user.permissions.length > 0) {
+            let ancestorIds = req.node.ancestorList.map(ancestor => ancestor._id);
+            
+            // Concat permission which have permissions about req.node
+            for (let permissionObject of user.permissions) {
+                // Check permission of upper node
+                if (Object.values(ancestorIds).some(ancestorId => ancestorId.equals(permissionObject._nodeId))) {
+                    // If have permission, add the permission to permissionList
+                    USER_GROUPS.forEach(userGroup => {
+                        if (userGroup._id.equals(permissionObject._userGroupId)) {
+                            userGroup.permissions.forEach(p => permissionList.add(p));
+                        }
+                    })
+                }
+            }
+        }
+    }
+
+    // Add default signupedUser rule
+    permissionList.add('defaultSignupedUser')
+
+    // Check permission rules
+    for (permission of permissionList) {
+        let filter = require('../../permissionRules/' + permission);
+        if(filter(req, user)) {
+            return true;
+        }
+    } 
+    return false;
+}
+
 
 module.exports = (req, res, next) => {
     let user = null;
